auth-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

sqlalchemy==2.0.25
pymysql==1.1.0

bcrypt==4.1.2

pyjwt==2.8.0
cryptography==42.0.5